@@ -10,31 +10,36 @@
         "build": "ng build --aot",
         "build:prod": "npm run configure -- --prod && npm run build -- --prod",
         "build:stats": "npm run configure -- --prod --no-mocks && npm run build -- --prod --stats-json --source-map",
-        "stats": "npm run build:stats && webpack-bundle-analyzer dist/brennus-tp/stats.json",
+        "stats:se": "cd dist/brennus-tp && cross-var source-map-explorer $npm_config_bundle ${npm_config_bundle}.map",
+        "stats:wba": "webpack-bundle-analyzer dist/brennus-tp/stats.json",
+        "stats": "npm run build:stats && npm run stats:wba",
         "lint": "ng lint --format stylish",
         "pretest": "npm run check:conf",
         "test": "ng test --karma-config src/karma-ci.conf.js --code-coverage",
         "test:watch": "ng test --karma-config src/karma.conf.js",
         "e2e": "ng e2e --port 4201",
         "prerelease": "npm ci && npm test && npm run e2e",
-        "release": "cross-var npm version $npm_package_new_version -m \"chore: Mark version v%s\"",
+        "release": "cross-var npm version $npm_config_new_version -m \"chore: Mark version v%s\"",
         "postrelease": "git push && git push --tags",
         "configure": "node src/conf/instrument.js",
         "check:conf": "node src/conf/check-conf.js",
-        "check:push": "npm ci && npm run build -- --prod && npm test && npm run e2e",
-        "compodoc": "compodoc -p src/tsconfig.app.json -d docs/compodoc -n 'Brennus Analytics'"
+        "check:push": "npm ci && npm run build:prod && npm test && npm run e2e",
+        "compodoc": "compodoc -p src/tsconfig.app.json -d docs/compodoc -n 'Brennus Analytics'",
+        "gen:svg-sprite": "gulp svg"
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "~7.0.0",
-        "@angular/common": "~7.0.0",
-        "@angular/compiler": "~7.0.0",
-        "@angular/core": "~7.0.0",
-        "@angular/forms": "~7.0.0",
-        "@angular/http": "~7.0.0",
-        "@angular/platform-browser": "~7.0.0",
-        "@angular/platform-browser-dynamic": "~7.0.0",
-        "@angular/router": "~7.0.0",
+        "@angular/animations": "^7.1.1",
+        "@angular/cdk": "^7.1.0",
+        "@angular/common": "~7.1.1",
+        "@angular/compiler": "~7.1.1",
+        "@angular/core": "~7.1.1",
+        "@angular/forms": "~7.1.1",
+        "@angular/http": "~7.1.1",
+        "@angular/material": "^7.1.0",
+        "@angular/platform-browser": "~7.1.1",
+        "@angular/platform-browser-dynamic": "~7.1.1",
+        "@angular/router": "~7.1.1",
         "@ngrx/effects": "^6.1.2",
         "@ngrx/entity": "^6.1.2",
         "@ngrx/router-store": "^6.1.2",
