@@ -1,25 +1,26 @@
 import { Component, ElementRef, OnDestroy, OnInit, ViewChild } from '@angular/core';
-import { GoodService } from '../good-service/good.service';
-import { fromEvent, merge, Observable, Subject } from 'rxjs';
-import { Good } from '../../core/model/good.model';
-import { ActivityIndicatorService } from '../../core/activity-indicator/activity-indicator.service';
+import { MatSlider } from '@angular/material';
 import { ActivatedRoute, Router } from '@angular/router';
-import { AppState } from '../../core/store/core.reducer';
 import { select, Store } from '@ngrx/store';
-import { SnackBar, SnackBarStategy, SnackBarType } from '../../core/snackbar/snackbar.service';
-import { goodSelectors } from '../store/good.selectors';
-import { debounceTime, distinctUntilKeyChanged, filter, first, map, startWith, takeUntil, tap } from 'rxjs/operators';
-import { BaHttpErrorResponse } from '../../core/network/ba-http-error-response';
-import { GSV_SetSearchFilters } from '../store/good.actions';
-import { SearchCriteria } from '../model/search-criteria.model';
 import { DatatableComponent } from '@swimlane/ngx-datatable';
+import { fromEvent, fromEventPattern, merge, Observable, Subject } from 'rxjs';
+import { debounceTime, distinctUntilKeyChanged, filter, first, map, startWith, takeUntil, tap } from 'rxjs/operators';
+import { ActivityIndicatorService } from '../../core/activity-indicator/activity-indicator.service';
+import { Good } from '../../core/model/good.model';
+import { BaHttpErrorResponse } from '../../core/network/ba-http-error-response';
+import { SnackBar } from '../../core/snackbar/snackbar.service';
+import { AppState } from '../../core/store/core.reducer';
+import { GoodService } from '../good-service/good.service';
+import { SearchCriteria } from '../model/search-criteria.model';
+import { GSV_SetSearchFilters } from '../store/good.actions';
+import { goodSelectors } from '../store/good.selectors';
 import { ColumnDefinition } from './column-definition';
 
 
 @Component({
     selector:'ba-good-search',
     templateUrl:'./good-search.component.html',
-    styleUrls:[ './good-search.component.scss' ]
+    styleUrls:['./good-search.component.scss']
 })
 export class GoodSearchComponent implements OnInit, OnDestroy {
 
@@ -27,9 +28,8 @@ export class GoodSearchComponent implements OnInit, OnDestroy {
     // -----------
 
     // DOM references
-    @ViewChild('searchInput', { read:ElementRef }) searchInput:ElementRef;
-    @ViewChild('minQualitySlider', { read:ElementRef }) minQualitySlider:ElementRef;
-    @ViewChild('searchForm', { read:ElementRef }) searchForm:ElementRef;
+    @ViewChild('searchInput', {read:ElementRef}) searchInput:ElementRef;
+    @ViewChild('minQualitySlider') minQualitySlider:MatSlider;
     // Current selected filters
     criteria = new SearchCriteria();
     goods$:Observable<Good[]>;
@@ -50,12 +50,10 @@ export class GoodSearchComponent implements OnInit, OnDestroy {
     footerHeight = 50;
     pageLimit = 50;
     // DOM References
-    @ViewChild('goodsTableContainer', { read:ElementRef }) goodsTableContainer:ElementRef;
+    @ViewChild('goodsTableContainer', {read:ElementRef}) goodsTableContainer:ElementRef;
     @ViewChild('datatable') datatable:DatatableComponent;
     // Columns definition
     columns:ColumnDefinition[];
-    // Observable fueled by the slider change events
-    private onMinQualityChange = new Subject<number>();
 
 
     // Utils
@@ -70,14 +68,14 @@ export class GoodSearchComponent implements OnInit, OnDestroy {
         private store:Store<AppState>,
         private router:Router,
         private route:ActivatedRoute,
-        private snackbar:SnackBar) {}
+        private snackbar:SnackBar) {
+    }
 
     ngOnInit() {
         this.listenToErrors();
         this.listenToSearchEvents();
         this.setColumns();
         this.setGoods();
-        this.initializeSliders();
         this.readCurrentCriteriaFromStore();
     }
 
@@ -108,24 +106,24 @@ export class GoodSearchComponent implements OnInit, OnDestroy {
             takeUntil(this.done)
         ).subscribe(
             (err:BaHttpErrorResponse) => {
-                this.snackbar.show({
-                    content:'Server is temporarily unavailable',
-                    type:SnackBarType.Error,
-                    strategy:SnackBarStategy.FirstOne
-                });
+                this.snackbar.showError('Server is temporarily unavailable');
             }
         );
     }
 
     listenToSearchEvents() {
-        const onFormSubmission = fromEvent(this.searchForm.nativeElement, 'submit', { passive:true }).pipe(
-            tap((event:Event) => {
-                event.preventDefault();
-                console.log('Form submitted');
+        const onKeywordsChange = fromEvent(this.searchInput.nativeElement, 'input', {passive:true});
+        const onKeywordsClear = fromEvent(this.searchInput.nativeElement, 'blur', {passive:true});
+
+        const onMinQualityChange = this.minQualitySlider.change.asObservable().pipe(
+            tap(event => {
+                const minQuality = event.value || 0;
+                console.log('minQuality:', minQuality);
+                this.criteria.minQuality = minQuality;
             })
         );
 
-        merge(onFormSubmission, this.onMinQualityChange).pipe(
+        merge(onKeywordsChange, onKeywordsClear, onMinQualityChange).pipe(
             debounceTime(500),
             map(() => this.criteria.updateHash().clone()),
             distinctUntilKeyChanged('hash'),
@@ -155,45 +153,22 @@ export class GoodSearchComponent implements OnInit, OnDestroy {
     setColumns() {
         if (!this.columns) {
             this.columns = [
-                { id:'name', name:'Name', minWidth:500, prop:'name' },
-                { id:'quality', name:'Quality', minWidth:150, prop:'quality' },
-                { id:'sellIn', name:'Time left (days)', minWidth:200, prop:'sellIn' },
-                { id:'type', name:'Type', minWidth:300, prop:'type' },
+                {id:'name', name:'Name', minWidth:500, prop:'name'},
+                {id:'quality', name:'Quality', minWidth:150, prop:'quality'},
+                {id:'sellIn', name:'Time left (days)', minWidth:200, prop:'sellIn'},
+                {id:'type', name:'Type', minWidth:300, prop:'type'},
             ];
         }
     }
 
     storeFilters(criteria:SearchCriteria) {
-        this.store.dispatch(new GSV_SetSearchFilters({ criteria:criteria.clone() }));
+        this.store.dispatch(new GSV_SetSearchFilters({criteria:criteria.clone()}));
         console.log('Search criteria updated in store');
     }
 
     clearSearchInput() {
         this.criteria.keywords = '';
         this.searchInput.nativeElement.focus();
-    }
-
-    initializeSliders() {
-        setTimeout(() => {
-            const slider = noUiSlider.create(this.minQualitySlider.nativeElement, {
-                start:[ 0 ],
-                connect:true,
-                step:1,
-                orientation:'horizontal',
-                range:{
-                    min:0,
-                    max:50
-                },
-                format:wNumb({
-                    decimals:0
-                })
-            });
-            this.minQualitySlider.nativeElement.noUiSlider.on('change', ([ minQuality ]) => {
-                minQuality = !minQuality ? 0 : parseInt(minQuality, 10);
-                this.criteria.minQuality = minQuality;
-                this.onMinQualityChange.next(minQuality);
-            });
-        }, 0);
     }
 
 }
